Move maybeLoadFromExternalFile to file.ts (#3228) - Moved maybeLoadFromExternalFile function from index.ts to file.ts - Added re-export in index.ts - Moved corresponding tests from index.test.ts to file.test.ts - Fixed maybeLoadToolsFromExternalFile to use the moved function - All tests passing
