+++ src/util/file.ts
@@ -1,6 +1,10 @@
-/**
- * Array of supported JavaScript and TypeScript file extensions
- */
+import { parse as csvParse } from 'csv-parse/sync';
+import yaml from 'js-yaml';
+import * as fs from 'fs';
+import * as path from 'path';
+import { getNunjucksEngine } from './templates';
+import cliState from '../cliState';
+
 export const JAVASCRIPT_EXTENSIONS = ['js', 'cjs', 'mjs', 'ts', 'cts', 'mts'];
 
 /**
@@ -47,4 +51,57 @@ export function isAudioFile(filePath: string): boolean {
   const audioExtensions = ['wav', 'mp3', 'ogg', 'aac', 'm4a', 'flac', 'wma', 'aiff', 'opus'];
   const fileExtension = filePath.split('.').pop()?.toLowerCase() || '';
   return audioExtensions.includes(fileExtension);
+}
+
+/**
+ * Loads content from an external file if the input is a file path, otherwise
+ * returns the input as-is. Supports Nunjucks templating for file paths.
+ *
+ * @param filePath - The input to process. Can be a file path string starting with "file://",
+ * an array of file paths, or any other type of data.
+ * @returns The loaded content if the input was a file path, otherwise the original input.
+ * For JSON and YAML files, the content is parsed into an object.
+ * For other file types, the raw file content is returned as a string.
+ *
+ * @throws {Error} If the specified file does not exist.
+ */
+export function maybeLoadFromExternalFile(filePath: string | object | Function | undefined | null) {
+  if (Array.isArray(filePath)) {
+    return filePath.map((path) => {
+      const content: any = maybeLoadFromExternalFile(path);
+      return content;
+    });
+  }
+
+  if (typeof filePath !== 'string') {
+    return filePath;
+  }
+  if (!filePath.startsWith('file://')) {
+    return filePath;
+  }
+
+  // Render the file path using Nunjucks
+  const renderedFilePath = getNunjucksEngine().renderString(filePath, {});
+
+  const finalPath = path.resolve(cliState.basePath || '', renderedFilePath.slice('file://'.length));
+  if (!fs.existsSync(finalPath)) {
+    throw new Error(`File does not exist: ${finalPath}`);
+  }
+
+  const contents = fs.readFileSync(finalPath, 'utf8');
+  if (finalPath.endsWith('.json')) {
+    return JSON.parse(contents);
+  }
+  if (finalPath.endsWith('.yaml') || finalPath.endsWith('.yml')) {
+    return yaml.load(contents);
+  }
+  if (finalPath.endsWith('.csv')) {
+    const records = csvParse(contents, { columns: true });
+    // If single column, return array of values
+    if (records.length > 0 && Object.keys(records[0]).length === 1) {
+      return records.map((record: Record<string, string>) => Object.values(record)[0]);
+    }
+    return records;
+  }
+  return contents;
 }
