+++ src/util/index.ts
@@ -451,58 +451,7 @@ export function parsePathOrGlob(
   };
 }
 
-/**
- * Loads content from an external file if the input is a file path, otherwise
- * returns the input as-is. Supports Nunjucks templating for file paths.
- *
- * @param filePath - The input to process. Can be a file path string starting with "file://",
- * an array of file paths, or any other type of data.
- * @returns The loaded content if the input was a file path, otherwise the original input.
- * For JSON and YAML files, the content is parsed into an object.
- * For other file types, the raw file content is returned as a string.
- *
- * @throws {Error} If the specified file does not exist.
- */
-export function maybeLoadFromExternalFile(filePath: string | object | Function | undefined | null) {
-  if (Array.isArray(filePath)) {
-    return filePath.map((path) => {
-      const content: any = maybeLoadFromExternalFile(path);
-      return content;
-    });
-  }
-
-  if (typeof filePath !== 'string') {
-    return filePath;
-  }
-  if (!filePath.startsWith('file://')) {
-    return filePath;
-  }
-
-  // Render the file path using Nunjucks
-  const renderedFilePath = getNunjucksEngine().renderString(filePath, {});
-
-  const finalPath = path.resolve(cliState.basePath || '', renderedFilePath.slice('file://'.length));
-  if (!fs.existsSync(finalPath)) {
-    throw new Error(`File does not exist: ${finalPath}`);
-  }
-
-  const contents = fs.readFileSync(finalPath, 'utf8');
-  if (finalPath.endsWith('.json')) {
-    return JSON.parse(contents);
-  }
-  if (finalPath.endsWith('.yaml') || finalPath.endsWith('.yml')) {
-    return yaml.load(contents);
-  }
-  if (finalPath.endsWith('.csv')) {
-    const records = csvParse(contents, { columns: true });
-    // If single column, return array of values
-    if (records.length > 0 && Object.keys(records[0]).length === 1) {
-      return records.map((record: Record<string, string>) => Object.values(record)[0]);
-    }
-    return records;
-  }
-  return contents;
-}
+export { maybeLoadFromExternalFile } from './file';
 
 export function isRunningUnderNpx(): boolean {
   const npmExecPath = getEnvString('npm_execpath');
@@ -528,5 +477,6 @@ export function maybeLoadToolsFromExternalFile(
   tools: any,
   vars?: Record<string, string | object>,
 ): any {
+  const { maybeLoadFromExternalFile } = require('./file');
   return maybeLoadFromExternalFile(renderVarsInObject(tools, vars));
 }
